+++ stories/Content/Tag/RemovableTag/RemovableTagWithIcon.stories.js
@@ -1,10 +1,10 @@
 import React from "react";
 
-import { TAG_COLORS } from "../../confetti-ds/src/constants";
-import { Button, RemovableTag as Component } from "../../confetti-ds/src";
+import { ICON_TYPES, TAG_COLORS } from "../../../../confetti-ds/src/constants";
+import { Button, RemovableTag as Component } from "../../../../confetti-ds/src";
 
 export default {
-  title: "Tag/Removable Tag With Thumb",
+  title: "Content/Tag/Removable Tag",
   component: Component,
   argTypes: {
     color: {
@@ -14,7 +14,7 @@ export default {
       control: { type: "inline-radio" },
     },
     icon: {
-      control: false,
+      control: { type: "select", options: ICON_TYPES },
     },
     thumbSrc: {
       control: false,
@@ -22,7 +22,7 @@ export default {
   },
 };
 
-export const RemovableTagWithThumb = (args) => {
+export const RemovableTagWithIcon = (args) => {
   const [showTag, setShowTag] = React.useState(true);
 
   return (
@@ -36,7 +36,7 @@ export const RemovableTagWithThumb = (args) => {
     </React.Fragment>
   )
 }
-RemovableTagWithThumb.args = {
+RemovableTagWithIcon.args = {
   text: "demo tag",
-  thumbSrc: "http://avatars3.githubusercontent.com/u/1887591",
+  icon: "coin",
 };
