+++ stories/Content/Tag/SimpleTag/SimpleTag.stories.js
@@ -4,7 +4,7 @@ import { TAG_COLORS } from "../../../../confetti-ds/src/constants";
 import { SimpleTag as Component } from "../../../../confetti-ds/src";
 
 export default {
-  title: "Content/Tag/Simple Tag/Simple Tag",
+  title: "Content/Tag/Simple Tag",
   component: Component,
   argTypes: {
     color: {
