+++ stories/Content/Tag/SimpleTag/SimpleTagWithIcon.stories.js
@@ -1,10 +1,10 @@
 import React from "react";
 
-import { TAG_COLORS } from "../../confetti-ds/src/constants";
-import { SimpleTag as Component } from "../../confetti-ds/src";
+import { ICON_TYPES, TAG_COLORS } from "../../../../confetti-ds/src/constants";
+import { SimpleTag as Component } from "../../../../confetti-ds/src";
 
 export default {
-  title: "Tag/Simple Tag With Thumb",
+  title: "Content/Tag/Simple Tag",
   component: Component,
   argTypes: {
     color: {
@@ -14,7 +14,7 @@ export default {
       control: { type: "inline-radio" },
     },
     icon: {
-      control: false, 
+      control: { type: "select", options: ICON_TYPES },
     },
     thumbSrc: {
       control: false,
@@ -22,7 +22,7 @@ export default {
   },
 };
 
-export const SimpleTagWithThumb = (args) => {
+export const SimpleTagWithIcon = (args) => {
   return (
     <React.Fragment>
       <Component {...args} />
@@ -30,7 +30,7 @@ export const SimpleTagWithThumb = (args) => {
   )
 };
 
-SimpleTagWithThumb.args = {
+SimpleTagWithIcon.args = {
   text: "demo tag",
-  thumbSrc: "http://avatars3.githubusercontent.com/u/1887591",
+  icon: "coin",
 };
